+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) as implemented by Cargo.
 
+## [unreleased] - 2020-11-29
+
+### Added
+
+- add support for M1/aarch64-apple-darwin/arm64-osx
+
 ## [0.2.10] - 2020-06-10
 
 ### Added